Add libdrop_ambient
This library can be used via LD_PRELOAD to force an application started with
ambient capabilities to drop them. It leaves other capabilities intact. This
can also be linked against and automatically does the right thing. You do
not need to make any calls into the library because all the work is done in
the constructor which runs before main() is called.