bindings/python: Allow hardcoded path to capability.h to be overridden
Currently the path to capability.h is hardcoded. When cross compiling
the host capabiity.h may be different to the target copy, leading
to different options being encoded in the python bindings than
expected. This causes a reproducibility issue amongst other potential
problems.

Add a configure option to optionally specify the right path to the
correct header as its probably safer/more reliable than trying to
query the compiler to get the header path.

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>